refactor(RacingGame): change category
