+++ src/test/java/racingcar/domain/RacingGameTest.java
@@ -7,7 +7,6 @@ import java.util.List;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import racingcar.MockRandomNumberGenerator;
-import racingcar.RacingGame;
 import racingcar.exception.GetWinnersBeforeFinishException;
 import racingcar.exception.RacingGameIsFinishedException;
 import racingcar.domain.result.MidtermResult;
